# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Chronus Booking API",
    "version": "1.0.0",
    "description": "API documentation for the Chronus multi-tenant mentor booking application."
  },
  "servers": [
    {
      "url": "http://localhost:3010/api/v1",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/bookings": {
      "get": {
        "summary": "List User Bookings",
        "description": "Fetches all bookings (active and cancelled) for the authenticated user as a member. Requires the session cookie.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of user bookings with slots and mentor details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "status": { "type": "string", "enum": ["ACTIVE", "CANCELLED"] },
                      "idempotencyKey": { "type": "string" },
                      "createdAt": { "type": "string", "format": "date-time" },
                      "updatedAt": { "type": "string", "format": "date-time" },
                      "slot": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "startTime": { "type": "string", "format": "date-time" },
                          "endTime": { "type": "string", "format": "date-time" },
                          "mentor": {
                            "type": "object",
                            "properties": {
                              "membershipId": { "type": "string", "format": "uuid" },
                              "userId": { "type": "string", "format": "uuid" },
                              "name": { "type": "string" },
                              "email": { "type": "string", "format": "email" },
                              "timezone": { "type": "string" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid session"
          },
          "403": {
            "description": "Access denied due to invalid membership"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create Booking",
        "description": "Book a mentor's availability slot. Employs a database transaction to verify availability and lock the slot to prevent concurrency/double-bookings. Idempotent key ensures request duplicate protection.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["slotId", "idempotencyKey"],
                "properties": {
                  "slotId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "9a38eb6c-7e61-46ab-844c-9f88d7456d98"
                  },
                  "idempotencyKey": {
                    "type": "string",
                    "example": "idempotent-key-12345"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "format": "uuid" },
                    "status": { "type": "string", "example": "ACTIVE" },
                    "idempotencyKey": { "type": "string" },
                    "createdAt": { "type": "string", "format": "date-time" },
                    "updatedAt": { "type": "string", "format": "date-time" },
                    "slot": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid" },
                        "startTime": { "type": "string", "format": "date-time" },
                        "endTime": { "type": "string", "format": "date-time" },
                        "mentor": {
                          "type": "object",
                          "properties": {
                            "membershipId": { "type": "string", "format": "uuid" },
                            "userId": { "type": "string", "format": "uuid" },
                            "name": { "type": "string" },
                            "email": { "type": "string", "format": "email" },
                            "timezone": { "type": "string" }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "200": {
            "description": "Booking replayed via idempotency key",
            "headers": {
              "x-idempotent-replayed": {
                "schema": {
                  "type": "string",
                  "example": "true"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input formatting or attempting to book own slot"
          },
          "401": {
            "description": "Authentication required or invalid session"
          },
          "403": {
            "description": "Access denied due to invalid membership"
          },
          "404": {
            "description": "Slot not found in organization"
          },
          "409": {
            "description": "Conflict. Slot is already booked or is no longer available"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/bookings/{bookingId}/cancel": {
      "post": {
        "summary": "Cancel Booking",
        "description": "Cancels an active booking and frees the slot. Uses Idempotency-Key if provided.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking cancelled successfully"
          },
          "400": {
            "description": "Cannot cancel a booking in the past"
          },
          "404": {
            "description": "Booking not found"
          }
        }
      }
    },
    "/bookings/{bookingId}/reschedule": {
      "post": {
        "summary": "Reschedule Booking",
        "description": "Reschedules an active booking to a new slot. Uses Idempotency-Key.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["newSlotId"],
                "properties": {
                  "newSlotId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking rescheduled successfully"
          },
          "400": {
            "description": "Invalid slot or booking status"
          },
          "404": {
            "description": "Booking or slot not found"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    }
  }
}